// UI/src/types/traffic.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import Entity from 'cs2/utils';

export enum TempFlags {
	Create = 1,
	Delete = 2,
	IsLast = 4,
	Essential = 8,
	Dragging = 16,
	Select = 32,
	Modify = 64,
	Regenerate = 128,
	Replace = 256,
	Upgrade = 512,
	Hidden = 1024,
	Parent = 2048,
	Combine = 4096,
	RemoveCost = 8192,
	Optional = 16384,
	Cancel = 32768,
	SubDetail = 65536
}
export interface float3
{
	x: number;
	y: number;
	z: number;
}
export interface DebugData
{
	entity: Entity;
	position: float3;
	position2d: float3;
	value: string;
	isEdge: boolean;
	isTemp: boolean;
	flags: TempFlags;
	original: Entity;
	originalIsEdge: boolean;
}